refactor: Remove dependency on group struct internals.
Also I don't know why one of the tests segfaults. This code is all kinds
of horrible. I'll investigate later.
